Refactor, Fix nomenclature, extend tests
Major refactor: standardised naming across libs (Text, Date, Array, File, Controls, Math, Clipboard)
added unit tests, removed dead code.
Resolved missing dependencies after DOF extraction (found Debug - Compile)

+++ Modules/libDBADO.bas
@@ -130,7 +130,7 @@ Public Function QuickValueSingle(sQuery As String, sReturnField As String) As Va
         Exit Function
     End If
 
-    Dim iCount As Integer: iCount = 0
+    Dim iCount As Long: iCount = 0
     db_Records.MoveFirst
 
     Do While Not db_Records.EOF
